+++ apps/client/src/app/pages/portfolio/activities/import-activities-dialog/import-activities-dialog.component.ts
@@ -346,6 +346,7 @@ export class GfImportActivitiesDialogComponent implements OnDestroy {
                 isDryRun: true,
                 tags: content.tags
               });
+
             this.activities = activities;
             this.dataSource = new MatTableDataSource(activities.reverse());
             this.pageIndex = 0;
@@ -360,15 +361,18 @@ export class GfImportActivitiesDialogComponent implements OnDestroy {
           const content = fileContent.split('\n').slice(1);
 
           try {
-            const data = await this.importActivitiesService.importCsv({
-              fileContent,
-              isDryRun: true,
-              userAccounts: this.data.user.accounts
-            });
-            this.activities = data.activities;
-            this.dataSource = new MatTableDataSource(data.activities.reverse());
+            const { activities, assetProfiles } =
+              await this.importActivitiesService.importCsv({
+                fileContent,
+                isDryRun: true,
+                userAccounts: this.data.user.accounts
+              });
+
+            this.activities = activities;
+            this.assetProfiles = assetProfiles;
+            this.dataSource = new MatTableDataSource(activities.reverse());
             this.pageIndex = 0;
-            this.totalItems = data.activities.length;
+            this.totalItems = activities.length;
           } catch (error) {
             console.error(error);
             this.handleImportError({