+++ CHANGELOG.md
@@ -20,10 +20,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved the currency validation in the search functionality of the data provider service
 - Optimized the get quotes functionality by utilizing the asset profile resolutions in the _Financial Modeling Prep_ service
 - Extracted the footer to a component
+- Refactored the blog page component to standalone
 - Improved the portfolio calculator unit tests to load the user currency from the exported file
+- Improved the language localization for German (`de`)
 
 ### Fixed
 
+- Fixed an issue in the `csv` file import where custom asset profiles failed due to validation errors
+- Fixed an issue with the total buy and sell calculation in the summary related to activities in a custom currency
 - Respected the include indices flag in the search functionality of the _Financial Modeling Prep_ service
 - Fixed an issue where the scroll position was not restored when changing pages
 - Fixed the word wrap in the menus of the activities table component